feat: moving  to utils, as we will reuse this in various providers
